fix!: test results directory name depends on the current branch name
Fixes #65 , #66 and #68 by changing the output directory name based on whether the current solution has a git repository.
This removes the need to check if the current solution has a git repository in both `IUnitTests.UnitTests` and `IMutationTests.MutationTests` targets whilst still making it possible to separate outputs by branch.